feat: ability to unregister default deno command
