+++ src/shell.ts
@@ -1,4 +1,3 @@
-import { existsSync } from "@std/fs/exists";
 import * as path from "@std/path";
 import { RealEnvironment as DenoWhichRealEnvironment, which } from "which";
 import type { KillSignal } from "./command.ts";
@@ -12,15 +11,10 @@ import {
   PipeSequencePipe,
   type PipeWriter,
   type Reader,
-  type ShellPipeReaderKind,
   ShellPipeWriter,
-  type ShellPipeWriterKind,
 } from "./pipes.ts";
 import { type EnvChange, type ExecuteResult, getAbortedResult } from "./result.ts";
-import type { SpawnedChildProcess } from "./runtimes/process.common.ts";
-import { spawnCommand } from "./runtimes/process.deno.ts";
-
-const neverAbortedSignal = new AbortController().signal;
+import { createExecutableCommand } from "./commands/executable.ts";
 
 export interface SequentialList {
   items: SequentialListItem[];
@@ -938,16 +932,6 @@ async function executeSimpleCommand(command: SimpleCommand, parentContext: Conte
   return await executeCommandArgs(commandArgs, context);
 }
 
-function checkMapCwdNotExistsError(cwd: string, err: unknown) {
-  if ((err as any).code === "ENOENT" && !existsSync(cwd)) {
-    throw new Error(`Failed to launch command because the cwd does not exist (${cwd}).`, {
-      cause: err,
-    });
-  } else {
-    throw err;
-  }
-}
-
 function executeCommandArgs(commandArgs: string[], context: Context): Promise<ExecuteResult> {
   // look for a registered command first
   const commandName = commandArgs.shift()!;
@@ -978,151 +962,13 @@ async function executeUnresolvedCommand(
     return executeUnresolvedCommand(resolvedCommand.command, [...resolvedCommand.args, ...commandArgs], context);
   }
   const _assertIsPath: "path" = resolvedCommand.kind;
-  return executeCommandAtPath(resolvedCommand.path, commandArgs, context);
-}
-
-async function executeCommandAtPath(
-  commandPath: string,
-  commandArgs: string[],
-  context: Context,
-): Promise<ExecuteResult> {
-  const pipeStringVals = {
-    stdin: getStdioStringValue(context.stdin),
-    stdout: getStdioStringValue(context.stdout.kind),
-    stderr: getStdioStringValue(context.stderr.kind),
-  };
-  let p: SpawnedChildProcess;
-  const cwd = context.getCwd();
-  try {
-    p = spawnCommand(commandPath, {
-      args: commandArgs,
-      cwd,
-      env: context.getEnvVars(),
-      clearEnv: true,
-      ...pipeStringVals,
-    });
-  } catch (err) {
-    // Deno throws this sync, Node.js throws it async
-    throw checkMapCwdNotExistsError(cwd, err);
-  }
-  const listener = (signal: Deno.Signal) => p.kill(signal);
-  context.signal.addListener(listener);
-  const completeController = new AbortController();
-  const completeSignal = completeController.signal;
-  let stdinError: unknown | undefined;
-  const stdinPromise = writeStdin(context.stdin, p, completeSignal)
-    .catch(async (err) => {
-      // don't surface anything because it's already been aborted
-      if (completeSignal.aborted) {
-        return;
-      }
-
-      const maybePromise = context.stderr.writeLine(`stdin pipe broken. ${errorToString(err)}`);
-      if (maybePromise != null) {
-        await maybePromise;
-      }
-      stdinError = err;
-      // kill the sub process
-      try {
-        p.kill("SIGKILL");
-      } catch (err) {
-        if (!(err instanceof Deno.errors.PermissionDenied || err instanceof Deno.errors.NotFound)) {
-          throw err;
-        }
-      }
-    });
-  try {
-    // don't abort stdout and stderr reads... ensure all of stdout/stderr is
-    // read in case the process exits before this finishes
-    const readStdoutTask = pipeStringVals.stdout === "piped"
-      ? readStdOutOrErr(p.stdout(), context.stdout)
-      : Promise.resolve();
-    const readStderrTask = pipeStringVals.stderr === "piped"
-      ? readStdOutOrErr(p.stderr(), context.stderr)
-      : Promise.resolve();
-    const [exitCode] = await Promise.all([
-      p.waitExitCode()
-        // for node.js, which throws this async
-        .catch((err) => Promise.reject(checkMapCwdNotExistsError(cwd, err))),
-      readStdoutTask,
-      readStderrTask,
-    ]);
-    if (stdinError != null) {
-      return {
-        code: 1,
-        kind: "exit",
-      };
-    } else {
-      return { code: exitCode };
-    }
-  } finally {
-    completeController.abort();
-    context.signal.removeListener(listener);
-
-    // ensure this is done before exiting... it will never throw
-    await stdinPromise;
-  }
-
-  async function writeStdin(stdin: CommandPipeReader, p: SpawnedChildProcess, signal: AbortSignal) {
-    if (typeof stdin === "string") {
-      return;
-    }
-    const processStdin = p.stdin();
-    await pipeReaderToWritable(stdin, processStdin, signal);
-    try {
-      await processStdin.close();
-    } catch {
-      // ignore
-    }
-  }
-
-  async function readStdOutOrErr(readable: ReadableStream<Uint8Array>, writer: ShellPipeWriter) {
-    if (typeof writer === "string") {
-      return;
-    }
-    // don't abort... ensure all of stdout/stderr is read in case the process
-    // exits before this finishes
-    await pipeReadableToWriterSync(readable, writer, neverAbortedSignal);
-  }
-
-  function getStdioStringValue(value: ShellPipeReaderKind | ShellPipeWriterKind) {
-    if (value === "inheritPiped") {
-      return "piped";
-    } else if (value === "inherit" || value === "null" || value === "piped") {
-      return value;
-    } else {
-      return "piped";
-    }
-  }
+  return createExecutableCommand(resolvedCommand.path)(context.asCommandContext(commandArgs));
 }
 
 async function executeSubshell(subshell: Subshell, context: Context): Promise<ExecuteResult> {
   const result = await executeSequentialList(subshell, context);
   // sub shells do not change the environment or cause an exit
   return { code: result.code };
-}
-
-async function pipeReaderToWritable(reader: Reader, writable: WritableStream<Uint8Array>, signal: AbortSignal) {
-  const abortedPromise = new Promise<void>((resolve) => {
-    signal.addEventListener("abort", listener);
-    function listener() {
-      signal.removeEventListener("abort", listener);
-      resolve();
-    }
-  });
-  const writer = writable.getWriter();
-  try {
-    while (!signal.aborted) {
-      const buffer = new Uint8Array(1024);
-      const length = await Promise.race([abortedPromise, reader.read(buffer)]);
-      if (length === 0 || length == null) {
-        break;
-      }
-      await writer.write(buffer.subarray(0, length));
-    }
-  } finally {
-    await writer.close();
-  }
 }
 
 async function pipeReaderToWriterSync(
@@ -1232,10 +1078,6 @@ export const denoWhichRealEnv = new WhichEnv();
 export async function whichFromContext(commandName: string, context: {
   getVar(key: string): string | undefined;
 }) {
-  // always use the current executable for "deno"
-  if (commandName.toUpperCase() === "DENO") {
-    return Deno.execPath();
-  }
   return await which(commandName, {
     os: Deno.build.os,
     stat: denoWhichRealEnv.stat,
